Database migrations and safe DB access
By default, sails uses the strategy `migrate:alter` on each
model to update the model in the database to represent any changes.
This can have detrimental effects when running in production:
data can be dropped, deleted, munged, etc.

Since there's no clear way to migrate schema changes, implement
manual database migrations.  By default, the sails strategy is
`migrate:safe` which will not make any changes to the database.

A very annoying aspect of sails is that the migration strategy
must be set for each MODEL, not for the system as a whole
or a specific adapter.  This means that each model had to be
modified to include the `migrate` parameter, pulled in from the
configuration file.  All new modules must have a `migrate`
parameter to keep consistent migration across all modules.

Use `node-db-migrate` to specify the changes to the database
when the models change.

INSTALL.md has been updated to explain how to run `db-migrate`
to get your database to the right migration version.

Inside `config/local.js`, there's a new variable `migrate`.
By default it is set to `safe` but can be reset to `alter` for
development (thus bypassing the database migration architecture)

The initial migration is also in this commit, which will bring
the database up to the current schema (for postgres).  Running
`db-migrate down` will also revert the database back to an
empty state.